Move the post-processing logic back
