--- battdat/postprocess/__init__.py
@@ -1,1 +1,0 @@
-"""Methods that compute new data based on existing columns in a cycling data file"""
