+++ battdat/postprocess/base.py
